+++ ChangeLog.md
@@ -12,6 +12,7 @@
 * Don't require build plan to upload [#400](https://github.com/commercialhaskell/stack/issues/400)
 * Specifying test components only builds/runs those tests [#398](https://github.com/commercialhaskell/stack/issues/398)
 * `STACK_EXE` environment variable
+* Add the `stack dot` command
 
 Bug fixes: